+++ CHANGELOG.md
@@ -1,6 +1,15 @@
-## 3.0.0 (unreleased)
+## 2.10.0 (Unreleased)
 
+* Added functions to retrieve and use VCD version `client.GetVcdVersion`, `client.GetVcdShortVersion`, `client.GetVcdFullVersion`, `client.VersionEqualOrGreater` [#339](https://github.com/vmware/go-vcloud-director/pull/339)
+* Added methods `VM.UpdateStorageProfile`, `VM.UpdateStorageProfileAsync` [#338](https://github.com/vmware/go-vcloud-director/pull/338)
+* Added methods `adminVdc.UpdateStorageProfile` [#340](https://github.com/vmware/go-vcloud-director/pull/340)
+* Added transparent support for bearer tokens [#341](https://github.com/vmware/go-vcloud-director/pull/341)
+* Added transparent connection using `cloudapi/1.0.0/sessions` when access through `api/sessions` is disabled
 * Able to upload an OVF without ovf:size defined in File part. Some bug fix for uploading OVA/OVF. [#331](https://github.com/vmware/go-vcloud-director/pull/331)
+
+BREAKING CHANGES:
+
+* type.VdcConfiguration (used for creation) changed the type for storage profile from `[]*VdcStorageProfile` to `[]*VdcStorageProfileConfiguration`
 
 ## 2.9.0 (October 15, 2020)